+++ tests/test_products.py
@@ -22,9 +22,10 @@ ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
+import copy
 from subprocess import CalledProcessError
 import unittest
-from unittest.mock import Mock, patch
+from unittest.mock import call, Mock, patch
 
 from yaml import safe_dump
 
@@ -35,24 +36,50 @@ from sat_install_utility.products import (
 )
 
 
+MOCK_PRODUCT_CATALOG_DATA = {
+    # Two versions of SAT that have no images in common with one another.
+    'sat': safe_dump({
+        '2.0.0': {'component_versions': {'docker': [
+            {'name': 'cray/cray-sat', 'version': '1.0.0'},
+            {'name': 'cray/sat-cfs-install', 'version': '1.4.0'}
+        ]}},
+        '2.0.1': {'component_versions': {'docker': [
+            {'name': 'cray/cray-sat', 'version': '1.0.1'},
+            {'name': 'cray/sat-other-image', 'version': '1.4.0'}
+        ]}},
+    }),
+    # Two versions of COS, where one of the images is the same between them.
+    'cos': safe_dump({
+        '2.0.0': {'component_versions': {'docker': [
+            {'name': 'cray/cray-cos', 'version': '1.0.0'},
+            {'name': 'cray/cos-cfs-install', 'version': '1.4.0'}
+        ]}},
+        '2.0.1': {'component_versions': {'docker': [
+            {'name': 'cray/cray-cos', 'version': '1.0.1'},
+            {'name': 'cray/cos-cfs-install', 'version': '1.4.0'}
+        ]}},
+    }),
+    # One version of "Other Product" that also uses cray/cray-sat:1.0.1
+    'other_product': safe_dump({
+        '2.0.0': {'component_versions': {'docker': [
+            {'name': 'cray/cray-sat', 'version': '1.0.1'},
+        ]}}
+    })
+}
+
+
 class TestProductCatalog(unittest.TestCase):
     """Tests for the ProductCatalog class."""
+
     def setUp(self):
         """Set up mocks."""
         self.mock_k8s_api = Mock()
-        self.mock_product_catalog_data = {
-            'sat': safe_dump({
-                '2.0.0': {'component_versions': {'sat': '1.0.0', 'sat-subpackage': '1.4.0'}},
-                '2.0.1': {'component_versions': {'sat': '1.0.1', 'sat-subpackage': '1.4.0'}},
-                '2.0.2': {'component_versions': {'sat': '1.0.1', 'sat-subpackage': '1.4.1'}},
-            }),
-            'cos': safe_dump({
-                '2.0.0': {'component_versions': {'cos': '1.0.0', 'cos-subpackage': '1.4.0'}},
-                '2.0.1': {'component_versions': {'cos': '1.0.1', 'cos-subpackage': '1.4.0'}},
-                '2.0.2': {'component_versions': {'cos': '1.0.1', 'cos-subpackage': '1.4.1'}},
-            })
-        }
+        self.mock_product_catalog_data = copy.deepcopy(MOCK_PRODUCT_CATALOG_DATA)
         self.mock_k8s_api.read_namespaced_config_map.return_value = Mock(data=self.mock_product_catalog_data)
+        self.mock_environ = patch('sat_install_utility.products.os.environ').start()
+        self.mock_check_output = patch('sat_install_utility.products.subprocess.check_output').start()
+        self.mock_print = patch('builtins.print').start()
+        self.mock_docker = Mock()
 
     def create_and_assert_product_catalog(self):
         """Assert the product catalog was created as expected."""
@@ -64,8 +91,8 @@ class TestProductCatalog(unittest.TestCase):
         """Test creating a simple ProductCatalog."""
         product_catalog = self.create_and_assert_product_catalog()
         expected_names_and_versions = [
-            (name, version) for name in ('sat', 'cos') for version in ('2.0.0', '2.0.1', '2.0.2')
-        ]
+            (name, version) for name in ('sat', 'cos') for version in ('2.0.0', '2.0.1')
+        ] + [('other_product', '2.0.0')]
         actual_names_and_versions = [
             (product.name, product.version) for product in product_catalog.products
         ]
@@ -90,16 +117,72 @@ class TestProductCatalog(unittest.TestCase):
         expected_matching_name_and_version = ('cos', '2.0.0')
         expected_other_versions = [('sat', version) for version in ('2.0.0', '2.0.1', '2.0.2')]
         expected_other_versions.extend([('cos', version) for version in ('2.0.1', '2.0.2')])
-        actual_matching_product, actual_other_versions = product_catalog.get_matching_products('cos', '2.0.0')
+        actual_matching_product = product_catalog.get_product('cos', '2.0.0')
         self.assertEqual(
             expected_matching_name_and_version, (actual_matching_product.name, actual_matching_product.version)
         )
-        self.assertEqual(
-            [(name, version) for name, version in expected_other_versions],
-            [(product.name, product.version) for product in actual_other_versions]
-        )
-        expected_component_data = {'component_versions': {'cos': '1.0.0', 'cos-subpackage': '1.4.0'}}
+        expected_component_data = {'component_versions': {'docker': [
+            {'name': 'cray/cray-cos', 'version': '1.0.0'},
+            {'name': 'cray/cos-cfs-install', 'version': '1.4.0'}
+        ]}}
         self.assertEqual(expected_component_data, actual_matching_product.data)
+
+    def test_remove_from_product_catalog(self):
+        """Test removing a version from the product catalog."""
+        product_catalog = self.create_and_assert_product_catalog()
+        product_catalog.remove_product_entry('mock_name', 'mock_version')
+        self.mock_environ.update.assert_called_once_with({
+            'PRODUCT': 'mock_name',
+            'PRODUCT_VERSION': 'mock_version',
+            'CONFIG_MAP': 'mock-name',
+            'CONFIG_MAP_NS': 'mock-namespace',
+            'VALIDATE_SCHEMA': 'true'
+        })
+        self.mock_check_output.assert_called_once_with(['catalog_delete.py'])
+
+    def test_remove_from_product_catalog_fail(self):
+        """Test removing a version from the product catalog when the subcommand fails."""
+        product_catalog = self.create_and_assert_product_catalog()
+        expected_err_regex = (
+            f'Error removing mock_name-mock_version from product catalog'
+        )
+        self.mock_check_output.side_effect = CalledProcessError(1, 'catalog_delete.py')
+        with self.assertRaisesRegex(ProductInstallException, expected_err_regex):
+            product_catalog.remove_product_entry('mock_name', 'mock_version')
+
+    def test_remove_product_docker_images(self):
+        """Test a basic removal of a product's docker images."""
+        product_catalog = self.create_and_assert_product_catalog()
+        with patch.object(InstalledProductVersion, 'uninstall_docker_image') as mock_uninstall_docker_image:
+            product_catalog.remove_product_docker_images('sat', '2.0.0', self.mock_docker)
+            mock_uninstall_docker_image.assert_has_calls([
+                call('cray/cray-sat', '1.0.0', self.mock_docker),
+                call('cray/sat-cfs-install', '1.4.0', self.mock_docker)
+            ])
+
+    def test_partial_remove_docker_images(self):
+        """Test a removal of docker images when an image is shared."""
+        product_catalog = self.create_and_assert_product_catalog()
+        with patch.object(InstalledProductVersion, 'uninstall_docker_image') as mock_uninstall_docker_image:
+            product_catalog.remove_product_docker_images('cos', '2.0.0', self.mock_docker)
+            mock_uninstall_docker_image.assert_called_once_with(
+                'cray/cray-cos', '1.0.0', self.mock_docker
+            )
+            self.mock_print.assert_called_once_with(
+                'Not removing Docker image cray/cos-cfs-install:1.4.0 '
+                'used by the following other product versions: cos-2.0.1'
+            )
+
+    def test_partial_remove_docker_images_other_product(self):
+        """Test a removal when a product's only image is also used by another product."""
+        product_catalog = self.create_and_assert_product_catalog()
+        with patch.object(InstalledProductVersion, 'uninstall_docker_image') as mock_uninstall_docker_image:
+            product_catalog.remove_product_docker_images('other_product', '2.0.0', self.mock_docker)
+            mock_uninstall_docker_image.assert_not_called()
+            self.mock_print.assert_called_once_with(
+                'Not removing Docker image cray/cray-sat:1.0.1 used '
+                'by the following other product versions: sat-2.0.1'
+            )
 
 
 class TestInstalledProductVersion(unittest.TestCase):
@@ -108,18 +191,10 @@ class TestInstalledProductVersion(unittest.TestCase):
         self.installed_product_version = InstalledProductVersion(
             'sat',
             '2.2.0',
-            {
-                'component_versions': {
-                    'docker': {
-                        'cray/cray-sat': '3.9.0',
-                        'cray/sat-cfs-install': '1.2.0',
-                        'cray/sat-install-utility': '3.9.0',
-                    },
-                    'rpm': {
-                        'cray-sat-podman': '1.5.0'
-                    }
-                }
-            }
+            {'component_versions': {'docker': [
+                {'name': 'cray/cray-sat', 'version': '1.0.2'},
+                {'name': 'cray/sat-cfs-install', 'version': '1.1.1'}
+            ]}}
         )
         self.legacy_installed_product_version = InstalledProductVersion(
             'sat', '1.0.1', {'component_versions': {'sat': '1.0.0'}}
@@ -133,8 +208,6 @@ class TestInstalledProductVersion(unittest.TestCase):
         # hosted repo, but good enough for the test.
         self.mock_nexus_api.repos.list.return_value = [self.mock_group_repo]
         self.mock_docker_api = Mock()
-        self.mock_environ = patch('sat_install_utility.products.os.environ').start()
-        self.mock_check_output = patch('sat_install_utility.products.subprocess.check_output').start()
 
     def tearDown(self):
         """Stop patches."""
@@ -142,16 +215,15 @@ class TestInstalledProductVersion(unittest.TestCase):
 
     def test_docker_images(self):
         """Test getting the Docker images."""
-        expected_docker_image_versions = {'cray/cray-sat': '3.9.0',
-                                          'cray/sat-cfs-install': '1.2.0',
-                                          'cray/sat-install-utility': '3.9.0'}
+        expected_docker_image_versions = [('cray/cray-sat', '1.0.2'),
+                                          ('cray/sat-cfs-install', '1.1.1')]
         self.assertEqual(
             expected_docker_image_versions, self.installed_product_version.docker_images
         )
 
     def test_legacy_docker_images(self):
         """Test getting the Docker images from an 'old'-style product catalog entry."""
-        expected_docker_image_versions = {'cray/cray-sat': '1.0.0'}
+        expected_docker_image_versions = [('cray/cray-sat', '1.0.0')]
         self.assertEqual(
             expected_docker_image_versions, self.legacy_installed_product_version.docker_images
         )
@@ -161,21 +233,21 @@ class TestInstalledProductVersion(unittest.TestCase):
         product_with_no_docker_images = InstalledProductVersion(
             'sat', '0.9.9', {'component_versions': {'docker': {}}}
         )
-        self.assertEqual(product_with_no_docker_images.docker_images, {})
+        self.assertEqual(product_with_no_docker_images.docker_images, [])
 
     def test_no_docker_images_null(self):
         """Test a product that has None under the 'docker' key returns an empty dictionary."""
         product_with_no_docker_images = InstalledProductVersion(
             'sat', '0.9.9', {'component_versions': {'docker': None}}
         )
-        self.assertEqual(product_with_no_docker_images.docker_images, {})
+        self.assertEqual(product_with_no_docker_images.docker_images, [])
 
     def test_no_docker_images_empty_list(self):
         """Test a product that has an empty list under the 'docker' key returns an empty dictionary."""
         product_with_no_docker_images = InstalledProductVersion(
             'sat', '0.9.9', {'component_versions': {'docker': []}}
         )
-        self.assertEqual(product_with_no_docker_images.docker_images, {})
+        self.assertEqual(product_with_no_docker_images.docker_images, [])
 
     def test_str(self):
         """Test the string representation of InstalledProductVersion."""
@@ -218,27 +290,6 @@ class TestInstalledProductVersion(unittest.TestCase):
             self.mock_group_repo.storage.strict_content_type_validation,
             member_names=('sat-2.2.0-sle-15sp3',)
         )
-
-    def test_remove_from_product_catalog(self):
-        """Test removing a version from the product catalog."""
-        self.installed_product_version.remove_from_product_catalog('mock-name', 'mock-namespace')
-        self.mock_environ.update.assert_called_once_with({
-            'PRODUCT': self.installed_product_version.name,
-            'PRODUCT_VERSION': self.installed_product_version.version,
-            'CONFIG_MAP': 'mock-name',
-            'CONFIG_MAP_NS': 'mock-namespace'
-        })
-        self.mock_check_output.assert_called_once_with(['catalog_delete.py'])
-
-    def test_remove_from_product_catalog_fail(self):
-        """Test removing a version from the product catalog when the subcommand fails."""
-        expected_err_regex = (
-            f'Error removing {self.installed_product_version.name}-'
-            f'{self.installed_product_version.version} from product catalog'
-        )
-        self.mock_check_output.side_effect = CalledProcessError(1, 'catalog_delete.py')
-        with self.assertRaisesRegex(ProductInstallException, expected_err_regex):
-            self.installed_product_version.remove_from_product_catalog('mock-name', 'mock-namespace')
 
 
 if __name__ == '__main__':