+++ CHANGELOG.md
@@ -29,6 +29,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed the product positional argument.
+- Refactored code that interacts with the product catalog in preparation for
+  change to use a common install utility library.
+- Change expected product catalog data format.
 - Changed the expected format of data in the ``cray-product-catalog`` ConfigMap
   so that multiple Docker images can be associated with a single product
   version. Maintains backwards compatibility with previous ``cray-product-catalog``